refactor: rename `_abstract_documentation_parser.py` to `_abstract_docstring_parser.py`
