+++ src/library_analyzer/processing/api/docstring_parsing/__init__.py
@@ -1,6 +1,6 @@
 """Parsing docstrings into a common format."""
 
-from ._abstract_documentation_parser import AbstractDocstringParser
+from ._abstract_docstring_parser import AbstractDocstringParser
 from ._create_docstring_parser import create_docstring_parser
 from ._docstring_style import DocstringStyle
 from ._epydoc_parser import EpydocParser
